Branching from matter-switch-test-modular-profiles to create subdriver
Using a subdriver for modular profile code and also separate some
functionality into button-utils and common-utils so that it can be
across multiple files.
