+++ drivers/SmartThings/matter-switch/src/init.lua
@@ -12,12 +12,15 @@
 -- See the License for the specific language governing permissions and
 -- limitations under the License.
 
+local button_utils = require "button-utils"
 local capabilities = require "st.capabilities"
+local color_utils = require "color_utils"
+local common_utils = require "common-utils"
 local log = require "log"
 local clusters = require "st.matter.clusters"
 local im = require "st.matter.interaction_model"
 local MatterDriver = require "st.matter.driver"
-local lua_socket = require "socket"
+local modular_profiles_utils = require "modular-profiles-utils"
 local utils = require "st.utils"
 local device_lib = require "st.device"
 local embedded_cluster_utils = require "embedded-cluster-utils"
@@ -44,15 +47,6 @@ local SWITCH_LEVEL_LIGHTING_MIN = 1
 local CURRENT_HUESAT_ATTR_MIN = 0
 local CURRENT_HUESAT_ATTR_MAX = 254
 
--- COMPONENT_TO_ENDPOINT_MAP is here to preserve the endpoint mapping for
--- devices that were joined to this driver as MCD devices before the transition
--- to join switch devices as parent-child. This value will exist in the device
--- table for devices that joined prior to this transition, and is also used for
--- button devices that require component mapping.
-local COMPONENT_TO_ENDPOINT_MAP = "__component_to_endpoint_map"
-local SUPPORTED_COMPONENT_CAPABILITIES = "__supported_component_capabilities"
-local ENERGY_MANAGEMENT_ENDPOINT = "__energy_management_endpoint"
-local IS_PARENT_CHILD_DEVICE = "__is_parent_child_device"
 local COLOR_TEMP_BOUND_RECEIVED_KELVIN = "__colorTemp_bound_received_kelvin"
 local COLOR_TEMP_BOUND_RECEIVED_MIRED = "__colorTemp_bound_received_mired"
 local COLOR_TEMP_MIN = "__color_temp_min"
@@ -63,163 +57,12 @@ local LEVEL_MAX = "__level_max"
 local COLOR_MODE = "__color_mode"
 
 local updated_fields = {
-  { current_field_name = "__component_to_endpoint_map_button", updated_field_name = COMPONENT_TO_ENDPOINT_MAP },
+  { current_field_name = "__component_to_endpoint_map_button", updated_field_name = common_utils.COMPONENT_TO_ENDPOINT_MAP },
   { current_field_name = "__switch_intialized", updated_field_name = nil }
 }
 
 local HUE_SAT_COLOR_MODE = clusters.ColorControl.types.ColorMode.CURRENT_HUE_AND_CURRENT_SATURATION
 local X_Y_COLOR_MODE = clusters.ColorControl.types.ColorMode.CURRENTX_AND_CURRENTY
-
-local AGGREGATOR_DEVICE_TYPE_ID = 0x000E
-local ON_OFF_LIGHT_DEVICE_TYPE_ID = 0x0100
-local DIMMABLE_LIGHT_DEVICE_TYPE_ID = 0x0101
-local COLOR_TEMP_LIGHT_DEVICE_TYPE_ID = 0x010C
-local EXTENDED_COLOR_LIGHT_DEVICE_TYPE_ID = 0x010D
-local ON_OFF_PLUG_DEVICE_TYPE_ID = 0x010A
-local DIMMABLE_PLUG_DEVICE_TYPE_ID = 0x010B
-local ON_OFF_SWITCH_ID = 0x0103
-local ON_OFF_DIMMER_SWITCH_ID = 0x0104
-local ON_OFF_COLOR_DIMMER_SWITCH_ID = 0x0105
-local MOUNTED_ON_OFF_CONTROL_ID = 0x010F
-local MOUNTED_DIMMABLE_LOAD_CONTROL_ID = 0x0110
-local GENERIC_SWITCH_ID = 0x000F
-local ELECTRICAL_SENSOR_ID = 0x0510
-local device_type_profile_map = {
-  [ON_OFF_LIGHT_DEVICE_TYPE_ID] = "light-binary",
-  [DIMMABLE_LIGHT_DEVICE_TYPE_ID] = "light-level",
-  [COLOR_TEMP_LIGHT_DEVICE_TYPE_ID] = "light-level-colorTemperature",
-  [EXTENDED_COLOR_LIGHT_DEVICE_TYPE_ID] = "light-color-level",
-  [ON_OFF_PLUG_DEVICE_TYPE_ID] = "plug-binary",
-  [DIMMABLE_PLUG_DEVICE_TYPE_ID] = "plug-level",
-  [ON_OFF_SWITCH_ID] = "switch-binary",
-  [ON_OFF_DIMMER_SWITCH_ID] = "switch-level",
-  [ON_OFF_COLOR_DIMMER_SWITCH_ID] = "switch-color-level",
-  [MOUNTED_ON_OFF_CONTROL_ID] = "switch-binary",
-  [MOUNTED_DIMMABLE_LOAD_CONTROL_ID] = "switch-level",
-}
-
-local device_categories = {
-  BUTTON = "BUTTON",
-  LIGHT = "LIGHT",
-  PLUG = "PLUG",
-  SWITCH = "SWITCH"
-}
-
-local device_type_category_map = {
-  [ON_OFF_LIGHT_DEVICE_TYPE_ID] = device_categories.LIGHT,
-  [DIMMABLE_LIGHT_DEVICE_TYPE_ID] = device_categories.LIGHT,
-  [COLOR_TEMP_LIGHT_DEVICE_TYPE_ID] = device_categories.LIGHT,
-  [EXTENDED_COLOR_LIGHT_DEVICE_TYPE_ID] = device_categories.LIGHT,
-  [ON_OFF_PLUG_DEVICE_TYPE_ID] = device_categories.PLUG,
-  [DIMMABLE_PLUG_DEVICE_TYPE_ID] = device_categories.PLUG,
-  [ON_OFF_SWITCH_ID] = device_categories.SWITCH,
-  [ON_OFF_DIMMER_SWITCH_ID] = device_categories.SWITCH,
-  [ON_OFF_COLOR_DIMMER_SWITCH_ID] = device_categories.SWITCH,
-  [MOUNTED_ON_OFF_CONTROL_ID] = device_categories.SWITCH,
-  [MOUNTED_DIMMABLE_LOAD_CONTROL_ID] = device_categories.SWITCH,
-  [GENERIC_SWITCH_ID] = device_categories.BUTTON
-}
-
-local device_type_attribute_map = {
-  [ON_OFF_LIGHT_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff
-  },
-  [DIMMABLE_LIGHT_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel
-  },
-  [COLOR_TEMP_LIGHT_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel,
-    clusters.ColorControl.attributes.ColorTemperatureMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMaxMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMinMireds
-  },
-  [EXTENDED_COLOR_LIGHT_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel,
-    clusters.ColorControl.attributes.ColorTemperatureMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMaxMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMinMireds,
-    clusters.ColorControl.attributes.CurrentHue,
-    clusters.ColorControl.attributes.CurrentSaturation,
-    clusters.ColorControl.attributes.CurrentX,
-    clusters.ColorControl.attributes.CurrentY
-  },
-  [ON_OFF_PLUG_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff
-  },
-  [DIMMABLE_PLUG_DEVICE_TYPE_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel
-  },
-  [ON_OFF_SWITCH_ID] = {
-    clusters.OnOff.attributes.OnOff
-  },
-  [ON_OFF_DIMMER_SWITCH_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel
-  },
-  [ON_OFF_COLOR_DIMMER_SWITCH_ID] = {
-    clusters.OnOff.attributes.OnOff,
-    clusters.LevelControl.attributes.CurrentLevel,
-    clusters.LevelControl.attributes.MaxLevel,
-    clusters.LevelControl.attributes.MinLevel,
-    clusters.ColorControl.attributes.ColorTemperatureMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMaxMireds,
-    clusters.ColorControl.attributes.ColorTempPhysicalMinMireds,
-    clusters.ColorControl.attributes.CurrentHue,
-    clusters.ColorControl.attributes.CurrentSaturation,
-    clusters.ColorControl.attributes.CurrentX,
-    clusters.ColorControl.attributes.CurrentY
-  },
-  [GENERIC_SWITCH_ID] = {
-    clusters.PowerSource.attributes.BatPercentRemaining,
-    clusters.Switch.events.InitialPress,
-    clusters.Switch.events.LongPress,
-    clusters.Switch.events.ShortRelease,
-    clusters.Switch.events.MultiPressComplete
-  },
-  [ELECTRICAL_SENSOR_ID] = {
-    clusters.ElectricalPowerMeasurement.attributes.ActivePower,
-    clusters.ElectricalEnergyMeasurement.attributes.CumulativeEnergyImported,
-    clusters.ElectricalEnergyMeasurement.attributes.PeriodicEnergyImported
-  }
-}
-
-local child_device_profile_overrides_per_vendor_id = {
-  [0x1321] = {
-    { product_id = 0x000C, target_profile = "switch-binary", initial_profile = "plug-binary" },
-    { product_id = 0x000D, target_profile = "switch-binary", initial_profile = "plug-binary" },
-  },
-  [0x115F] = {
-    { product_id = 0x1003, target_profile = "light-power-energy-powerConsumption" },       -- 2 Buttons(Generic Switch), 1 Channel(On/Off Light)
-    { product_id = 0x1004, target_profile = "light-power-energy-powerConsumption" },       -- 2 Buttons(Generic Switch), 2 Channels(On/Off Light)
-    { product_id = 0x1005, target_profile = "light-power-energy-powerConsumption" },       -- 4 Buttons(Generic Switch), 3 Channels(On/Off Light)
-    { product_id = 0x1006, target_profile = "light-level-power-energy-powerConsumption" }, -- 3 Buttons(Generic Switch), 1 Channels(Dimmable Light)
-    { product_id = 0x1008, target_profile = "light-power-energy-powerConsumption" },       -- 2 Buttons(Generic Switch), 1 Channel(On/Off Light)
-    { product_id = 0x1009, target_profile = "light-power-energy-powerConsumption" },       -- 4 Buttons(Generic Switch), 2 Channels(On/Off Light)
-    { product_id = 0x100A, target_profile = "light-level-power-energy-powerConsumption" }, -- 1 Buttons(Generic Switch), 1 Channels(Dimmable Light)
-  }
-}
-
-local battery_support = {
-  NO_BATTERY = "NO_BATTERY",
-  BATTERY_LEVEL = "BATTERY_LEVEL",
-  BATTERY_PERCENTAGE = "BATTERY_PERCENTAGE"
-}
-
-local detect_matter_thing
 
 local CUMULATIVE_REPORTS_NOT_SUPPORTED = "__cumulative_reports_not_supported"
 local FIRST_IMPORT_REPORT_TIMESTAMP = "__first_import_report_timestamp"
@@ -260,7 +103,7 @@ local function send_import_poll_report(device, latest_total_imported_energy_wh)
   end
 
   -- Report the energy consumed during the time interval. The unit of these values should be 'Wh'
-  if not device:get_field(ENERGY_MANAGEMENT_ENDPOINT) then
+  if not device:get_field(common_utils.ENERGY_MANAGEMENT_ENDPOINT) then
     device:emit_event(capabilities.powerConsumptionReport.powerConsumption({
       start = iso8061Timestamp(last_time),
       ["end"] = iso8061Timestamp(current_time - 1),
@@ -268,7 +111,7 @@ local function send_import_poll_report(device, latest_total_imported_energy_wh)
       energy = latest_total_imported_energy_wh
     }))
   else
-    device:emit_event_for_endpoint(device:get_field(ENERGY_MANAGEMENT_ENDPOINT),capabilities.powerConsumptionReport.powerConsumption({
+    device:emit_event_for_endpoint(device:get_field(common_utils.ENERGY_MANAGEMENT_ENDPOINT),capabilities.powerConsumptionReport.powerConsumption({
       start = iso8061Timestamp(last_time),
       ["end"] = iso8061Timestamp(current_time - 1),
       deltaEnergy = energy_delta_wh,
@@ -307,30 +150,12 @@ local function set_poll_report_timer_and_schedule(device, is_cumulative_report)
     -- the poll schedule is only needed for devices that support powerConsumption
     -- and enable powerConsumption when energy management is defined in root endpoint(0).
     if device:supports_capability(capabilities.powerConsumptionReport) or
-       device:get_field(ENERGY_MANAGEMENT_ENDPOINT) then
+       device:get_field(common_utils.ENERGY_MANAGEMENT_ENDPOINT) then
       create_poll_report_schedule(device)
     end
     device:set_field(IMPORT_POLL_TIMER_SETTING_ATTEMPTED, true)
   end
 end
-
-local START_BUTTON_PRESS = "__start_button_press"
-local TIMEOUT_THRESHOLD = 10 --arbitrary timeout
-local HELD_THRESHOLD = 1
--- this is the number of buttons for which we have a static profile already made
-local STATIC_BUTTON_PROFILE_SUPPORTED = {1, 2, 3, 4, 5, 6, 7, 8}
-
--- Some switches will send a MultiPressComplete event as part of a long press sequence. Normally the driver will create a
--- button capability event on receipt of MultiPressComplete, but in this case that would result in an extra event because
--- the "held" capability event is generated when the LongPress event is received. The IGNORE_NEXT_MPC flag is used
--- to tell the driver to ignore MultiPressComplete if it is received after a long press to avoid this extra event.
-local IGNORE_NEXT_MPC = "__ignore_next_mpc"
-
--- These are essentially storing the supported features of a given endpoint
--- TODO: add an is_feature_supported_for_endpoint function to matter.device that takes an endpoint
-local EMULATE_HELD = "__emulate_held" -- for non-MSR (MomentarySwitchRelease) devices we can emulate this on the software side
-local SUPPORTS_MULTI_PRESS = "__multi_button" -- for MSM devices (MomentarySwitchMultiPress), create an event on receipt of MultiPressComplete
-local INITIAL_PRESS_ONLY = "__initial_press_only" -- for devices that support MS (MomentarySwitch), but not MSR (MomentarySwitchRelease)
 
 local TEMP_BOUND_RECEIVED = "__temp_bound_received"
 local TEMP_MIN = "__temp_min"
@@ -338,57 +163,6 @@ local TEMP_MAX = "__temp_max"
 
 local AQARA_MANUFACTURER_ID = 0x115F
 local AQARA_CLIMATE_SENSOR_W100_ID = 0x2004
-
-local function supports_modular_profile(device)
-  return version.api >= 14 and version.rpc >= 8 and
-    not (device.manufacturer_info.vendor_id == AQARA_MANUFACTURER_ID and
-         device.manufacturer_info.product_id == AQARA_CLIMATE_SENSOR_W100_ID)
-end
-
---helper function to create list of multi press values
-local function create_multi_press_values_list(size, supportsHeld)
-  local list = {"pushed", "double"}
-  if supportsHeld then table.insert(list, "held") end
-  -- add multi press values of 3 or greater to the list
-  for i=3, size do
-    table.insert(list, string.format("pushed_%dx", i))
-  end
-  return list
-end
-
-local function tbl_contains(array, value)
-  for _, element in ipairs(array) do
-    if element == value then
-      return true
-    end
-  end
-  return false
-end
-
-local function get_field_for_endpoint(device, field, endpoint)
-  return device:get_field(string.format("%s_%d", field, endpoint))
-end
-
-local function set_field_for_endpoint(device, field, endpoint, value, additional_params)
-  device:set_field(string.format("%s_%d", field, endpoint), value, additional_params)
-end
-
-local function init_press(device, endpoint)
-  set_field_for_endpoint(device, START_BUTTON_PRESS, endpoint, lua_socket.gettime(), {persist = false})
-end
-
-local function emulate_held_event(device, ep)
-  local now = lua_socket.gettime()
-  local press_init = get_field_for_endpoint(device, START_BUTTON_PRESS, ep) or now -- if we don't have an init time, assume instant release
-  if (now - press_init) < TIMEOUT_THRESHOLD then
-    if (now - press_init) > HELD_THRESHOLD then
-      device:emit_event_for_endpoint(ep, capabilities.button.button.held({state_change = true}))
-    else
-      device:emit_event_for_endpoint(ep, capabilities.button.button.pushed({state_change = true}))
-    end
-  end
-  set_field_for_endpoint(device, START_BUTTON_PRESS, ep, nil, {persist = false})
-end
 
 local function convert_huesat_st_to_matter(val)
   return utils.clamp_value(math.floor((val * 0xFE) / 100.0 + 0.5), CURRENT_HUESAT_ATTR_MIN, CURRENT_HUESAT_ATTR_MAX)
@@ -416,94 +190,6 @@ local function mired_to_kelvin(value, minOrMax)
   end
 end
 
---- device_type_supports_button_switch_combination helper function used to check
---- whether the device type for an endpoint is currently supported by a profile for
---- combination button/switch devices.
-local function device_type_supports_button_switch_combination(device, endpoint_id)
-  for _, ep in ipairs(device.endpoints) do
-    if ep.endpoint_id == endpoint_id then
-      for _, dt in ipairs(ep.device_types) do
-        if dt.device_type_id == DIMMABLE_LIGHT_DEVICE_TYPE_ID then
-          for _, fingerprint in ipairs(child_device_profile_overrides_per_vendor_id[0x115F]) do
-            if device.manufacturer_info.product_id == fingerprint.product_id then
-              return false -- For Aqara Dimmer Switch with Button.
-            end
-          end
-          return true
-        end
-      end
-    end
-  end
-  return false
-end
-
-local function get_first_non_zero_endpoint(endpoints)
-  table.sort(endpoints)
-  for _,ep in ipairs(endpoints) do
-    if ep ~= 0 then -- 0 is the matter RootNode endpoint
-      return ep
-    end
-  end
-  return nil
-end
-
---- find_default_endpoint is a helper function to handle situations where
---- device does not have endpoint ids in sequential order from 1
-local function find_default_endpoint(device)
-  if device.manufacturer_info.vendor_id == AQARA_MANUFACTURER_ID and
-     device.manufacturer_info.product_id == AQARA_CLIMATE_SENSOR_W100_ID then
-    -- In case of Aqara Climate Sensor W100, in order to sequentially set the button name to button 1, 2, 3
-    return device.MATTER_DEFAULT_ENDPOINT
-  end
-
-  local switch_eps = device:get_endpoints(clusters.OnOff.ID)
-  local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
-
-  -- Return the first switch endpoint as the default endpoint if no button endpoints are present
-  if #button_eps == 0 and #switch_eps > 0 then
-    return get_first_non_zero_endpoint(switch_eps)
-  end
-
-  -- Return the first button endpoint as the default endpoint if no switch endpoints are present
-  if #switch_eps == 0 and #button_eps > 0 then
-    return get_first_non_zero_endpoint(button_eps)
-  end
-
-  -- If both switch and button endpoints are present, check the device type on the main switch
-  -- endpoint. If it is not a supported device type, return the first button endpoint as the
-  -- default endpoint.
-  if #switch_eps > 0 and #button_eps > 0 then
-    local main_endpoint = get_first_non_zero_endpoint(switch_eps)
-    if supports_modular_profile(device) or device_type_supports_button_switch_combination(device, main_endpoint) then
-      return get_first_non_zero_endpoint(switch_eps)
-    else
-      device.log.warn("The main switch endpoint does not contain a supported device type for a component configuration with buttons")
-      return get_first_non_zero_endpoint(button_eps)
-    end
-  end
-
-  device.log.warn(string.format("Did not find default endpoint, will use endpoint %d instead", device.MATTER_DEFAULT_ENDPOINT))
-  return device.MATTER_DEFAULT_ENDPOINT
-end
-
-local function component_to_endpoint(device, component)
-  local map = device:get_field(COMPONENT_TO_ENDPOINT_MAP) or {}
-  if map[component] then
-    return map[component]
-  end
-  return find_default_endpoint(device)
-end
-
-local function endpoint_to_component(device, ep)
-  local map = device:get_field(COMPONENT_TO_ENDPOINT_MAP) or {}
-  for component, endpoint in pairs(map) do
-    if endpoint == ep then
-      return component
-    end
-  end
-  return "main"
-end
-
 local function check_field_name_updates(device)
   for _, field in ipairs(updated_fields) do
     if device:get_field(field.current_field_name) then
@@ -515,467 +201,43 @@ local function check_field_name_updates(device)
   end
 end
 
-local function assign_child_profile(device, child_ep)
-  local profile
-
-  for _, ep in ipairs(device.endpoints) do
-    if ep.endpoint_id == child_ep then
-      -- Some devices report multiple device types which are a subset of
-      -- a superset device type (For example, Dimmable Light is a superset of
-      -- On/Off light). This mostly applies to the four light types, so we will want
-      -- to match the profile for the superset device type. This can be done by
-      -- matching to the device type with the highest ID
-      local id = 0
-      for _, dt in ipairs(ep.device_types) do
-        id = math.max(id, dt.device_type_id)
-      end
-      profile = device_type_profile_map[id]
-      break
-    end
-  end
-
-  -- Check if device has an overridden child profile that differs from the profile that would match
-  -- the child's device type for the following two cases:
-  --   1. To add Electrical Sensor only to the first EDGE_CHILD (light-power-energy-powerConsumption)
-  --      for the Aqara Light Switch H2. The profile of the second EDGE_CHILD for this device is
-  --      determined in the "for" loop above (e.g., light-binary)
-  --   2. The selected profile for the child device matches the initial profile defined in
-  --      child_device_profile_overrides
-  for id, vendor in pairs(child_device_profile_overrides_per_vendor_id) do
-    for _, fingerprint in ipairs(vendor) do
-      if device.manufacturer_info.product_id == fingerprint.product_id and
-         ((device.manufacturer_info.vendor_id == AQARA_MANUFACTURER_ID and child_ep == 1) or profile == fingerprint.initial_profile) then
-         return fingerprint.target_profile
-      end
-    end
-  end
-
-  -- default to "switch-binary" if no profile is found
-  return profile or "switch-binary"
-end
-
-local function configure_buttons(device)
-  local ms_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
-  local msr_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH_RELEASE})
-  local msl_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH_LONG_PRESS})
-  local msm_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH_MULTI_PRESS})
-
-  for _, ep in ipairs(ms_eps) do
-    if device.profile.components[endpoint_to_component(device, ep)] then
-      device.log.info_with({hub_logs=true}, string.format("Configuring Supported Values for generic switch endpoint %d", ep))
-      local supportedButtonValues_event
-      -- this ordering is important, since MSM & MSL devices must also support MSR
-      if tbl_contains(msm_eps, ep) then
-        supportedButtonValues_event = nil -- deferred to the max press handler
-        device:send(clusters.Switch.attributes.MultiPressMax:read(device, ep))
-        set_field_for_endpoint(device, SUPPORTS_MULTI_PRESS, ep, true, {persist = true})
-      elseif tbl_contains(msl_eps, ep) then
-        supportedButtonValues_event = capabilities.button.supportedButtonValues({"pushed", "held"}, {visibility = {displayed = false}})
-      elseif tbl_contains(msr_eps, ep) then
-        supportedButtonValues_event = capabilities.button.supportedButtonValues({"pushed", "held"}, {visibility = {displayed = false}})
-        set_field_for_endpoint(device, EMULATE_HELD, ep, true, {persist = true})
-      else -- this switch endpoint only supports momentary switch, no release events
-        supportedButtonValues_event = capabilities.button.supportedButtonValues({"pushed"}, {visibility = {displayed = false}})
-        set_field_for_endpoint(device, INITIAL_PRESS_ONLY, ep, true, {persist = true})
-      end
-
-      if supportedButtonValues_event then
-        device:emit_event_for_endpoint(ep, supportedButtonValues_event)
-      end
-      device:emit_event_for_endpoint(ep, capabilities.button.button.pushed({state_change = false}))
-    else
-      device.log.info_with({hub_logs=true}, string.format("Component not found for generic switch endpoint %d. Skipping Supported Value configuration", ep))
-    end
-  end
-end
-
 local function find_child(parent, ep_id)
   return parent:get_child_by_parent_assigned_key(string.format("%d", ep_id))
-end
-
-local function supports_capability_by_id_modular(device, capability, component)
-  if not device:get_field(SUPPORTED_COMPONENT_CAPABILITIES) then return false end
-  for _, component_capabilities in ipairs(device:get_field(SUPPORTED_COMPONENT_CAPABILITIES)) do
-    local comp_id = component_capabilities[1]
-    local capability_ids = component_capabilities[2]
-    if (component == nil) or (component == comp_id) then
-      for _, cap in ipairs(capability_ids) do
-        if cap == capability then
-          return true
-        end
-      end
-    end
-  end
-  return false
-end
-
-local function build_button_component_map(device, main_endpoint, button_eps)
-  -- create component mapping on the main profile button endpoints
-  table.sort(button_eps)
-  local component_map = {}
-  component_map["main"] = main_endpoint
-  for component_num, ep in ipairs(button_eps) do
-    if ep ~= main_endpoint then
-      local button_component = "button"
-      if #button_eps > 1 then
-        button_component = button_component .. component_num
-      end
-      component_map[button_component] = ep
-    end
-  end
-  device:set_field(COMPONENT_TO_ENDPOINT_MAP, component_map, {persist = true})
-end
-
-local function build_button_profile(device, main_endpoint, num_button_eps)
-  local profile_name = string.gsub(num_button_eps .. "-button", "1%-", "") -- remove the "1-" in a device with 1 button ep
-  if device_type_supports_button_switch_combination(device, main_endpoint) then
-    profile_name = "light-level-" .. profile_name
-  end
-  if #device:get_endpoints(clusters.PowerSource.ID, {feature_bitmap = clusters.PowerSource.types.PowerSourceFeature.BATTERY}) == 0 then
-    device:try_update_metadata({profile = profile_name})
-  else
-    device:send(clusters.PowerSource.attributes.AttributeList:read(device)) -- battery profiles are configured later, in power_source_attribute_list_handler
-  end
-end
-
-local function build_child_switch_profiles(driver, device, main_endpoint)
-  local num_switch_server_eps = 0
-  local parent_child_device = false
-  local switch_eps = device:get_endpoints(clusters.OnOff.ID)
-  table.sort(switch_eps)
-  for _, ep in ipairs(switch_eps) do
-    if device:supports_server_cluster(clusters.OnOff.ID, ep) then
-      num_switch_server_eps = num_switch_server_eps + 1
-      if ep ~= main_endpoint then -- don't create a child device that maps to the main endpoint
-        local name = string.format("%s %d", device.label, num_switch_server_eps)
-        local child_profile = assign_child_profile(device, ep)
-        driver:try_create_device(
-          {
-            type = "EDGE_CHILD",
-            label = name,
-            profile = child_profile,
-            parent_device_id = device.id,
-            parent_assigned_child_key = string.format("%d", ep),
-            vendor_provided_label = name
-          }
-        )
-        parent_child_device = true
-        if _ == 1 and string.find(child_profile, "energy") then
-          -- when energy management is defined in the root endpoint(0), replace it with the first switch endpoint and process it.
-          device:set_field(ENERGY_MANAGEMENT_ENDPOINT, ep, {persist = true})
-        end
-      end
-    end
-  end
-
-  -- If the device is a parent child device, set the find_child function on init. This is persisted because initialize_buttons_and_switches
-  -- is only run once, but find_child function should be set on each driver init.
-  if parent_child_device then
-    device:set_field(IS_PARENT_CHILD_DEVICE, true, {persist = true})
-  end
-
-  -- this is needed in initialize_buttons_and_switches
-  return num_switch_server_eps
-end
-
-local function handle_light_switch_with_onOff_server_clusters(device, main_endpoint, return_device_type)
-  local cluster_id = 0
-  for _, ep in ipairs(device.endpoints) do
-    -- main_endpoint only supports server cluster by definition of get_endpoints()
-    if main_endpoint == ep.endpoint_id then
-      for _, dt in ipairs(ep.device_types) do
-        -- no device type that is not in the switch subset should be considered.
-        if (ON_OFF_SWITCH_ID <= dt.device_type_id and dt.device_type_id <= ON_OFF_COLOR_DIMMER_SWITCH_ID) then
-          cluster_id = math.max(cluster_id, dt.device_type_id)
-        end
-      end
-      break
-    end
-  end
-
-  if return_device_type then
-    return cluster_id
-  end
-
-  if device_type_profile_map[cluster_id] then
-    device:try_update_metadata({profile = device_type_profile_map[cluster_id]})
-  end
 end
 
 local function initialize_buttons_and_switches(driver, device, main_endpoint)
   local profile_found = false
   local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
-  if tbl_contains(STATIC_BUTTON_PROFILE_SUPPORTED, #button_eps) then
-    build_button_component_map(device, main_endpoint, button_eps)
-    build_button_profile(device, main_endpoint, #button_eps)
+  if common_utils.tbl_contains(button_utils.STATIC_BUTTON_PROFILE_SUPPORTED, #button_eps) then
+    button_utils.build_button_component_map(device, main_endpoint, button_eps)
+    button_utils.build_button_profile(device, main_endpoint, #button_eps)
     -- All button endpoints found will be added as additional components in the profile containing the main_endpoint.
     -- The resulting endpoint to component map is saved in the COMPONENT_TO_ENDPOINT_MAP field
-    configure_buttons(device)
+    button_utils.configure_buttons(device)
     profile_found = true
   end
   -- Without support for bindings, only clusters that are implemented as server are counted. This count is handled
   -- while building switch child profiles
-  local num_switch_server_eps = build_child_switch_profiles(driver, device, main_endpoint)
+  local num_switch_server_eps = common_utils.build_child_switch_profiles(driver, device, main_endpoint)
   -- We do not support the Light Switch device types because they require OnOff to be implemented as 'client', which requires us to support bindings.
   -- However, this workaround profiles devices that claim to be Light Switches, but that break spec and implement OnOff as 'server'.
   -- Note: since their device type isn't supported, these devices join as a matter-thing.
-  if num_switch_server_eps > 0 and detect_matter_thing(device) then
-    handle_light_switch_with_onOff_server_clusters(device, main_endpoint)
+  if num_switch_server_eps > 0 and common_utils.detect_matter_thing(device) then
+    common_utils.handle_light_switch_with_onOff_server_clusters(device, main_endpoint)
     profile_found = true
   end
   return profile_found
 end
 
-local function detect_bridge(device)
-  for _, ep in ipairs(device.endpoints) do
-    for _, dt in ipairs(ep.device_types) do
-      if dt.device_type_id == AGGREGATOR_DEVICE_TYPE_ID then
-        return true
-      end
-    end
-  end
-  return false
-end
-
---- get_device_category helper function to determine the category that should be
---- used in a device's profile. The more specific categories are preferred,
---- except for Button, because buttons are included as optional components in
---- every modular profile. Order of preference:
----   1. Light / Plug / Water Valve
----   2. Switch
----   3. Button
-local function get_device_category(device)
-  local button_found = false
-  local switch_found = false
-  for _, ep in ipairs(device.endpoints) do
-    for _, dt in ipairs(ep.device_types) do
-      local category = device_type_category_map[dt.device_type_id]
-      if category == device_categories.LIGHT or category == device_categories.PLUG or category == device_categories.WATER_VALVE then
-        return category
-      elseif category == device_categories.SWITCH then
-        switch_found = true
-      elseif category == device_categories.BUTTON then
-        button_found = true
-      end
-    end
-  end
-  if switch_found then
-    return device_categories.SWITCH
-  end
-  if button_found then
-    return device_categories.BUTTON
-  end
-  -- Return SWITCH as default if no other category is found
-  return device_categories.SWITCH
-end
-
-local function add_battery_capability(component_capabilities, battery_attr_support)
-  if battery_attr_support == battery_support.BATTERY_PERCENTAGE then
-    table.insert(component_capabilities, capabilities.battery.ID)
-  elseif battery_attr_support == battery_support.BATTERY_LEVEL then
-    table.insert(component_capabilities, capabilities.batteryLevel.ID)
-  end
-end
-
-local function add_energy_and_power_capabilities(component_capabilities, num_energy_eps, num_power_eps)
-  if num_energy_eps > 0 and num_power_eps > 0 then
-    table.insert(component_capabilities, capabilities.powerMeter.ID)
-    table.insert(component_capabilities, capabilities.energyMeter.ID)
-    table.insert(component_capabilities, capabilities.powerConsumptionReport.ID)
-  elseif num_energy_eps > 0 then
-    table.insert(component_capabilities, capabilities.energyMeter.ID)
-    table.insert(component_capabilities, capabilities.powerConsumptionReport.ID)
-  elseif num_power_eps > 0 then
-    table.insert(component_capabilities, capabilities.powerMeter.ID)
-  end
-end
-
-local function match_modular_profile(driver, device, battery_attr_support)
-  local main_endpoint = find_default_endpoint(device)
-  local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap = clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
-  local color_hs_eps = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.types.Feature.HS})
-  local color_temp_eps = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.types.Feature.CT})
-  local color_xy_eps = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.types.Feature.XY})
-  local energy_eps = device:get_endpoints(clusters.ElectricalEnergyMeasurement.ID)
-  local fan_eps = device:get_endpoints(clusters.FanControl.ID)
-  local humidity_eps = device:get_endpoints(clusters.RelativeHumidityMeasurement.ID)
-  local level_eps = device:get_endpoints(clusters.LevelControl.ID)
-  local power_eps = device:get_endpoints(clusters.ElectricalPowerMeasurement.ID)
-  local switch_eps = device:get_endpoints(clusters.OnOff.ID)
-  local temperature_eps = device:get_endpoints(clusters.TemperatureMeasurement.ID)
-  local valve_eps = device:get_endpoints(clusters.ValveConfigurationAndControl.ID)
-
-  local category = get_device_category(device)
-
-  local optional_supported_component_capabilities = {}
-  local main_component_capabilities = {}
-  local extra_component_capabilities = {}
-
-  local MAIN_COMPONENT_IDX = 1
-  local CAPABILITIES_LIST_IDX = 2
-
-  if #button_eps > 0 then
-    for component_num, _ in ipairs(button_eps) do
-      -- button-modular profile uses 'main', 'button2', button3', ... as component names.
-      -- Other profiles use 'main', 'button', 'button2', etc
-      if component_num == 1 and category == device_categories.BUTTON then
-        table.insert(main_component_capabilities, capabilities.button.ID)
-        if battery_attr_support then
-          add_battery_capability(main_component_capabilities, battery_attr_support)
-        end
-      else
-        local button_capabilities = {}
-        table.insert(button_capabilities, capabilities.button.ID)
-        if component_num == 1 and battery_attr_support then
-          add_battery_capability(button_capabilities, battery_attr_support)
-        end
-        local component_name = "button"
-        if component_num > 1 then
-          component_name = component_name .. component_num
-        end
-        table.insert(extra_component_capabilities, {component_name, button_capabilities})
-      end
-    end
-    build_button_component_map(device, main_endpoint, button_eps)
-    configure_buttons(device)
-  end
-
-  -- Only add capabilities related to lights if the corresponding cluster is
-  -- implemented on the main endpoint. Otherwise, it will be added as a child device.
-  if tbl_contains(color_hs_eps, main_endpoint) or tbl_contains(color_xy_eps, main_endpoint) then
-    table.insert(main_component_capabilities, capabilities.colorControl.ID)
-  end
-
-  if tbl_contains(color_temp_eps, main_endpoint) then
-    table.insert(main_component_capabilities, capabilities.colorTemperature.ID)
-  end
-
-  if tbl_contains(level_eps, main_endpoint) then
-    table.insert(main_component_capabilities, capabilities.switchLevel.ID)
-  end
-
-  if #fan_eps > 0 then
-    table.insert(main_component_capabilities, capabilities.fanMode.ID)
-    table.insert(main_component_capabilities, capabilities.fanSpeedPercent.ID)
-  end
-
-  if #humidity_eps > 0 then
-    table.insert(main_component_capabilities, capabilities.relativeHumidityMeasurement.ID)
-  end
-
-  add_energy_and_power_capabilities(main_component_capabilities, #energy_eps, #power_eps)
-
-  if #switch_eps > 0 then
-    if category == device_categories.PLUG then
-      for component_num, ep in ipairs(switch_eps) do
-        if component_num ~= 1 then
-          local plug_capabilities = {}
-          table.insert(plug_capabilities, capabilities.switch.ID)
-          if tbl_contains(level_eps, ep) then
-            table.insert(plug_capabilities, capabilities.switchLevel.ID)
-          end
-          add_energy_and_power_capabilities(plug_capabilities, #energy_eps, #power_eps)
-          local component_name = "plug" .. component_num
-          table.insert(extra_component_capabilities, {component_name, plug_capabilities})
-        end
-      end
-    elseif category == device_categories.BUTTON or category == device_categories.WATER_VALVE then
-      table.insert(main_component_capabilities, capabilities.switch.ID)
-    else -- category = LIGHT or SWITCH
-      local num_switch_server_eps = build_child_switch_profiles(driver, device, main_endpoint)
-      if num_switch_server_eps > 0 and detect_matter_thing(device) then
-        -- Ensure that the proper capabilities are included for Light Switch
-        -- device types that implement the OnOff cluster as 'server'
-        local device_type_id = handle_light_switch_with_onOff_server_clusters(device, main_endpoint, true)
-        if ON_OFF_SWITCH_ID <= device_type_id and device_type_id <= ON_OFF_COLOR_DIMMER_SWITCH_ID then
-          local capabilities_to_remove = {}
-          if device_type_id == ON_OFF_SWITCH_ID then
-            capabilities_to_remove = {capabilities.colorControl.ID, capabilities.colorTemperature.ID, capabilities.switchLevel.ID}
-          elseif device_type_id == ON_OFF_DIMMER_SWITCH_ID then
-            capabilities_to_remove = {capabilities.colorControl.ID, capabilities.colorTemperature.ID}
-            if not tbl_contains(main_component_capabilities, capabilities.switchLevel.ID) then
-              table.insert(main_component_capabilities, capabilities.switchLevel.ID)
-            end
-          else -- device_type_id = ON_OFF_COLOR_DIMMER_SWITCH_ID
-            if not tbl_contains(main_component_capabilities, capabilities.switchLevel.ID) then
-              table.insert(main_component_capabilities, capabilities.switchLevel.ID)
-            end
-            if not tbl_contains(main_component_capabilities, capabilities.colorTemperature.ID) then
-              table.insert(main_component_capabilities, capabilities.colorTemperature.ID)
-            end
-            if not tbl_contains(main_component_capabilities, capabilities.colorControl.ID) then
-              table.insert(main_component_capabilities, capabilities.colorControl.ID)
-            end
-          end
-          for _, capability in ipairs(capabilities_to_remove) do
-            local _, found_idx = tbl_contains(main_component_capabilities, capability)
-            if found_idx then
-              table.remove(main_component_capabilities, found_idx)
-            end
-          end
-        end
-      end
-    end
-  end
-
-  if #temperature_eps > 0 then
-    table.insert(main_component_capabilities, capabilities.temperatureMeasurement.ID)
-  end
-
-  if #valve_eps > 0 then
-    table.insert(main_component_capabilities, capabilities.valve.ID)
-    if #embedded_cluster_utils.get_endpoints(device, clusters.ValveConfigurationAndControl.ID,
-      {feature_bitmap = clusters.ValveConfigurationAndControl.types.Feature.LEVEL}) > 0 then
-      table.insert(main_component_capabilities, capabilities.level.ID)
-    end
-  end
-
-  table.insert(optional_supported_component_capabilities, {"main", main_component_capabilities})
-  for _, component_capability in ipairs(extra_component_capabilities) do
-    table.insert(optional_supported_component_capabilities, component_capability)
-  end
-
-  local total_supported_capabilities = optional_supported_component_capabilities
-
-  local profile_name, mandatory_capability_for_device_category
-  if category == device_categories.BUTTON then
-    profile_name = "button-modular"
-    mandatory_capability_for_device_category = capabilities.button.ID
-  elseif category == device_categories.LIGHT then
-    profile_name = "light-modular"
-    mandatory_capability_for_device_category = capabilities.switch.ID
-  elseif category == device_categories.PLUG then
-    profile_name = "plug-modular"
-    mandatory_capability_for_device_category = capabilities.switch.ID
-  elseif category == device_categories.WATER_VALVE then
-    profile_name = "water-valve-modular"
-    mandatory_capability_for_device_category = capabilities.valve.ID
-  else -- category = device_categories.SWITCH
-    profile_name = "switch-modular"
-    mandatory_capability_for_device_category = capabilities.switch.ID
-  end
-
-  device:try_update_metadata({profile = profile_name, optional_component_capabilities = optional_supported_component_capabilities})
-
-  -- add mandatory capabilities for subscription
-  table.insert(total_supported_capabilities[MAIN_COMPONENT_IDX][CAPABILITIES_LIST_IDX], mandatory_capability_for_device_category)
-  table.insert(total_supported_capabilities[MAIN_COMPONENT_IDX][CAPABILITIES_LIST_IDX], capabilities.refresh.ID)
-  table.insert(total_supported_capabilities[MAIN_COMPONENT_IDX][CAPABILITIES_LIST_IDX], capabilities.firmwareUpdate.ID)
-
-  device:set_field(SUPPORTED_COMPONENT_CAPABILITIES, total_supported_capabilities, {persist = true})
-  -- re-up subscription with new capabilities using the modular supports_capability override
-  device:extend_device("supports_capability_by_id", supports_capability_by_id_modular)
-end
-
 local function device_init(driver, device)
   if device.network_type == device_lib.NETWORK_TYPE_MATTER then
     check_field_name_updates(device)
-    device:set_component_to_endpoint_fn(component_to_endpoint)
-    device:set_endpoint_to_component_fn(endpoint_to_component)
-    if device:get_field(IS_PARENT_CHILD_DEVICE) then
+    device:set_component_to_endpoint_fn(common_utils.component_to_endpoint)
+    device:set_endpoint_to_component_fn(common_utils.endpoint_to_component)
+    if device:get_field(common_utils.IS_PARENT_CHILD_DEVICE) then
       device:set_find_child(find_child)
     end
-    local main_endpoint = find_default_endpoint(device)
+    local main_endpoint = common_utils.find_default_endpoint(device)
     -- ensure subscription to all endpoint attributes- including those mapped to child devices
     for _, ep in ipairs(device.endpoints) do
       if ep.endpoint_id ~= main_endpoint then
@@ -983,8 +245,8 @@ local function device_init(driver, device)
         for _, dt in ipairs(ep.device_types) do
           id = math.max(id, dt.device_type_id)
         end
-        for _, attr in pairs(device_type_attribute_map[id] or {}) do
-          if id == GENERIC_SWITCH_ID and
+        for _, attr in pairs(common_utils.device_type_attribute_map[id] or {}) do
+          if id == common_utils.GENERIC_SWITCH_ID and
              attr ~= clusters.PowerSource.attributes.BatPercentRemaining and
              attr ~= clusters.PowerSource.attributes.BatChargeLevel then
             device:add_subscribed_event(attr)
@@ -994,69 +256,61 @@ local function device_init(driver, device)
         end
       end
     end
-    if device:get_field(SUPPORTED_COMPONENT_CAPABILITIES) then
-      device:extend_device("supports_capability_by_id", supports_capability_by_id_modular)
+    if device:get_field(modular_profiles_utils.SUPPORTED_COMPONENT_CAPABILITIES) then
+      device:extend_device("supports_capability_by_id", modular_profiles_utils.supports_capability_by_id_modular)
     end
     device:subscribe()
   end
 end
 
 local function match_profile(driver, device)
-  if supports_modular_profile(device) then
-    if #device:get_endpoints(clusters.PowerSource.ID, {feature_bitmap = clusters.PowerSource.types.PowerSourceFeature.BATTERY}) == 0 then
-      match_modular_profile(driver, device)
-    else
-      device:send(clusters.PowerSource.attributes.AttributeList:read(device)) -- battery profiles are configured later, in power_source_attribute_list_handler
+  local main_endpoint = common_utils.find_default_endpoint(device)
+  -- initialize the main device card with buttons if applicable, and create child devices as needed for multi-switch devices.
+  local profile_found = initialize_buttons_and_switches(driver, device, main_endpoint)
+  if device:get_field(common_utils.IS_PARENT_CHILD_DEVICE) then
+    device:set_find_child(find_child)
+  end
+  if profile_found then
+    return
+  end
+  local fan_eps = device:get_endpoints(clusters.FanControl.ID)
+  local level_eps = device:get_endpoints(clusters.LevelControl.ID)
+  local energy_eps = embedded_cluster_utils.get_endpoints(device, clusters.ElectricalEnergyMeasurement.ID)
+  local power_eps = embedded_cluster_utils.get_endpoints(device, clusters.ElectricalPowerMeasurement.ID)
+  local valve_eps = embedded_cluster_utils.get_endpoints(device, clusters.ValveConfigurationAndControl.ID)
+  local profile_name
+  local level_support = ""
+  if #level_eps > 0 then
+    level_support = "-level"
+  end
+  if #energy_eps > 0 and #power_eps > 0 then
+    profile_name = "plug" .. level_support .. "-power-energy-powerConsumption"
+  elseif #energy_eps > 0 then
+    profile_name = "plug" .. level_support .. "-energy-powerConsumption"
+  elseif #power_eps > 0 then
+    profile_name = "plug" .. level_support .. "-power"
+  elseif #valve_eps > 0 then
+    profile_name = "water-valve"
+    if #embedded_cluster_utils.get_endpoints(device, clusters.ValveConfigurationAndControl.ID,
+      {feature_bitmap = clusters.ValveConfigurationAndControl.types.Feature.LEVEL}) > 0 then
+      profile_name = profile_name .. "-level"
     end
-  else
-    local main_endpoint = find_default_endpoint(device)
-    -- initialize the main device card with buttons if applicable, and create child devices as needed for multi-switch devices.
-    local profile_found = initialize_buttons_and_switches(driver, device, main_endpoint)
-    if device:get_field(IS_PARENT_CHILD_DEVICE) then
-      device:set_find_child(find_child)
-    end
-    if profile_found then
-      return
-    end
-    local fan_eps = device:get_endpoints(clusters.FanControl.ID)
-    local level_eps = device:get_endpoints(clusters.LevelControl.ID)
-    local energy_eps = embedded_cluster_utils.get_endpoints(device, clusters.ElectricalEnergyMeasurement.ID)
-    local power_eps = embedded_cluster_utils.get_endpoints(device, clusters.ElectricalPowerMeasurement.ID)
-    local valve_eps = embedded_cluster_utils.get_endpoints(device, clusters.ValveConfigurationAndControl.ID)
-    local profile_name
-    local level_support = ""
-    if #level_eps > 0 then
-      level_support = "-level"
-    end
-    if #energy_eps > 0 and #power_eps > 0 then
-      profile_name = "plug" .. level_support .. "-power-energy-powerConsumption"
-    elseif #energy_eps > 0 then
-      profile_name = "plug" .. level_support .. "-energy-powerConsumption"
-    elseif #power_eps > 0 then
-      profile_name = "plug" .. level_support .. "-power"
-    elseif #valve_eps > 0 then
-      profile_name = "water-valve"
-      if #embedded_cluster_utils.get_endpoints(device, clusters.ValveConfigurationAndControl.ID,
-        {feature_bitmap = clusters.ValveConfigurationAndControl.types.Feature.LEVEL}) > 0 then
-        profile_name = profile_name .. "-level"
-      end
-    elseif #fan_eps > 0 then
-      profile_name = "light-color-level-fan"
-    end
-    if profile_name then
-      device:try_update_metadata({ profile = profile_name })
-    end
+  elseif #fan_eps > 0 then
+    profile_name = "light-color-level-fan"
+  end
+  if profile_name then
+    device:try_update_metadata({ profile = profile_name })
   end
 end
 
 local function do_configure(driver, device)
-  if device.network_type == device_lib.NETWORK_TYPE_MATTER and not detect_bridge(device) then
+  if device.network_type == device_lib.NETWORK_TYPE_MATTER and not common_utils.detect_bridge(device) then
     match_profile(driver, device)
   end
 end
 
 local function driver_switched(driver, device)
-  if device.network_type == device_lib.NETWORK_TYPE_MATTER and not detect_bridge(device) then
+  if device.network_type == device_lib.NETWORK_TYPE_MATTER and not common_utils.detect_bridge(device) then
     match_profile(driver, device)
   end
 end
@@ -1105,7 +359,7 @@ local function handle_set_color(driver, device, cmd)
   local endpoint_id = device:component_to_endpoint(cmd.component)
   local req
   local huesat_endpoints = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.FeatureMap.HUE_AND_SATURATION})
-  if tbl_contains(huesat_endpoints, endpoint_id) then
+  if common_utils.tbl_contains(huesat_endpoints, endpoint_id) then
     local hue = convert_huesat_st_to_matter(cmd.args.color.hue)
     local sat = convert_huesat_st_to_matter(cmd.args.color.saturation)
     req = clusters.ColorControl.server.commands.MoveToHueAndSaturation(device, endpoint_id, hue, sat, TRANSITION_TIME, OPTIONS_MASK, OPTIONS_OVERRIDE)
@@ -1119,7 +373,7 @@ end
 local function handle_set_hue(driver, device, cmd)
   local endpoint_id = device:component_to_endpoint(cmd.component)
   local huesat_endpoints = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.FeatureMap.HUE_AND_SATURATION})
-  if tbl_contains(huesat_endpoints, endpoint_id) then
+  if common_utils.tbl_contains(huesat_endpoints, endpoint_id) then
     local hue = convert_huesat_st_to_matter(cmd.args.hue)
     local req = clusters.ColorControl.server.commands.MoveToHue(device, endpoint_id, hue, 0, TRANSITION_TIME, OPTIONS_MASK, OPTIONS_OVERRIDE)
     device:send(req)
@@ -1131,7 +385,7 @@ end
 local function handle_set_saturation(driver, device, cmd)
   local endpoint_id = device:component_to_endpoint(cmd.component)
   local huesat_endpoints = device:get_endpoints(clusters.ColorControl.ID, {feature_bitmap = clusters.ColorControl.FeatureMap.HUE_AND_SATURATION})
-  if tbl_contains(huesat_endpoints, endpoint_id) then
+  if common_utils.tbl_contains(huesat_endpoints, endpoint_id) then
     local sat = convert_huesat_st_to_matter(cmd.args.saturation)
     local req = clusters.ColorControl.server.commands.MoveToSaturation(device, endpoint_id, sat, TRANSITION_TIME, OPTIONS_MASK, OPTIONS_OVERRIDE)
     device:send(req)
@@ -1143,14 +397,14 @@ end
 local function handle_set_color_temperature(driver, device, cmd)
   local endpoint_id = device:component_to_endpoint(cmd.component)
   local temp_in_kelvin = cmd.args.temperature
-  local min_temp_kelvin = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, endpoint_id)
-  local max_temp_kelvin = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, endpoint_id)
+  local min_temp_kelvin = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, endpoint_id)
+  local max_temp_kelvin = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, endpoint_id)
 
   local temp_in_mired = utils.round(MIRED_KELVIN_CONVERSION_CONSTANT/temp_in_kelvin)
   if min_temp_kelvin ~= nil and temp_in_kelvin <= min_temp_kelvin then
-    temp_in_mired = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, endpoint_id)
+    temp_in_mired = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, endpoint_id)
   elseif max_temp_kelvin ~= nil and temp_in_kelvin >= max_temp_kelvin then
-    temp_in_mired = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, endpoint_id)
+    temp_in_mired = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, endpoint_id)
   end
   local req = clusters.ColorControl.server.commands.MoveToColorTemperature(device, endpoint_id, temp_in_mired, TRANSITION_TIME, OPTIONS_MASK, OPTIONS_OVERRIDE)
   device:set_field(MOST_RECENT_TEMP, cmd.args.temperature, {persist = true})
@@ -1258,18 +512,18 @@ local function temp_attr_handler(driver, device, ib, response)
     device.log.warn_with({hub_logs = true}, string.format("Device reported color temperature %d mired outside of sane range of %.2f-%.2f", temp_in_mired, COLOR_TEMPERATURE_MIRED_MIN, COLOR_TEMPERATURE_MIRED_MAX))
     return
   end
-  local min_temp_mired = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, ib.endpoint_id)
-  local max_temp_mired = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, ib.endpoint_id)
+  local min_temp_mired = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, ib.endpoint_id)
+  local max_temp_mired = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, ib.endpoint_id)
 
   local temp = utils.round(MIRED_KELVIN_CONVERSION_CONSTANT/temp_in_mired)
   if min_temp_mired ~= nil and temp_in_mired <= min_temp_mired then
-    temp = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, ib.endpoint_id)
+    temp = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, ib.endpoint_id)
   elseif max_temp_mired ~= nil and temp_in_mired >= max_temp_mired then
-    temp = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, ib.endpoint_id)
+    temp = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, ib.endpoint_id)
   end
 
   local temp_device = device
-  if device:get_field(IS_PARENT_CHILD_DEVICE) == true then
+  if device:get_field(common_utils.IS_PARENT_CHILD_DEVICE) == true then
     temp_device = find_child(device, ib.endpoint_id) or device
   end
   local most_recent_temp = temp_device:get_field(MOST_RECENT_TEMP)
@@ -1293,15 +547,15 @@ local mired_bounds_handler_factory = function(minOrMax)
       return
     end
     local temp_in_kelvin = mired_to_kelvin(temp_in_mired, minOrMax)
-    set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..minOrMax, ib.endpoint_id, temp_in_kelvin)
+    common_utils.set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..minOrMax, ib.endpoint_id, temp_in_kelvin)
     -- the minimum color temp in kelvin corresponds to the maximum temp in mireds
     if minOrMax == COLOR_TEMP_MIN then
-      set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, ib.endpoint_id, temp_in_mired)
+      common_utils.set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MAX, ib.endpoint_id, temp_in_mired)
     else
-      set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, ib.endpoint_id, temp_in_mired)
+      common_utils.set_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_MIRED..COLOR_TEMP_MIN, ib.endpoint_id, temp_in_mired)
     end
-    local min = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, ib.endpoint_id)
-    local max = get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, ib.endpoint_id)
+    local min = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MIN, ib.endpoint_id)
+    local max = common_utils.get_field_for_endpoint(device, COLOR_TEMP_BOUND_RECEIVED_KELVIN..COLOR_TEMP_MAX, ib.endpoint_id)
     if min ~= nil and max ~= nil then
       if min < max then
         device:emit_event_for_endpoint(ib.endpoint_id, capabilities.colorTemperature.colorTemperatureRange({ value = {minimum = min, maximum = max} }))
@@ -1318,7 +572,7 @@ local level_bounds_handler_factory = function(minOrMax)
       return
     end
     local lighting_endpoints = device:get_endpoints(clusters.LevelControl.ID, {feature_bitmap = clusters.LevelControl.FeatureMap.LIGHTING})
-    local lighting_support = tbl_contains(lighting_endpoints, ib.endpoint_id)
+    local lighting_support = common_utils.tbl_contains(lighting_endpoints, ib.endpoint_id)
     -- If the lighting feature is supported then we should check if the reported level is at least 1.
     if lighting_support and ib.data.value < SWITCH_LEVEL_LIGHTING_MIN then
       device.log.warn_with({hub_logs = true}, string.format("Lighting device reported a switch level %d outside of supported capability range", ib.data.value))
@@ -1330,22 +584,20 @@ local level_bounds_handler_factory = function(minOrMax)
     if lighting_support and level == 0 then
       level = 1
     end
-    set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..minOrMax, ib.endpoint_id, level)
-    local min = get_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MIN, ib.endpoint_id)
-    local max = get_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MAX, ib.endpoint_id)
+    common_utils.set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..minOrMax, ib.endpoint_id, level)
+    local min = common_utils.get_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MIN, ib.endpoint_id)
+    local max = common_utils.get_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MAX, ib.endpoint_id)
     if min ~= nil and max ~= nil then
       if min < max then
         device:emit_event_for_endpoint(ib.endpoint_id, capabilities.switchLevel.levelRange({ value = {minimum = min, maximum = max} }))
       else
         device.log.warn_with({hub_logs = true}, string.format("Device reported a min level value %d that is not lower than the reported max level value %d", min, max))
       end
-      set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MAX, ib.endpoint_id, nil)
-      set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MIN, ib.endpoint_id, nil)
+      common_utils.set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MAX, ib.endpoint_id, nil)
+      common_utils.set_field_for_endpoint(device, LEVEL_BOUND_RECEIVED..LEVEL_MIN, ib.endpoint_id, nil)
     end
   end
 end
-
-local color_utils = require "color_utils"
 
 local function x_attr_handler(driver, device, ib, response)
   if device:get_field(COLOR_MODE) == HUE_SAT_COLOR_MODE then
@@ -1425,7 +677,7 @@ local function cumul_energy_imported_handler(driver, device, ib, response)
       device:emit_event_for_endpoint(ib.endpoint_id, capabilities.energyMeter.energy({ value = watt_hour_value, unit = "Wh" }))
     else
       -- when energy management is defined in the root endpoint(0), replace it with the first switch endpoint and process it.
-      device:emit_event_for_endpoint(device:get_field(ENERGY_MANAGEMENT_ENDPOINT), capabilities.energyMeter.energy({ value = watt_hour_value, unit = "Wh" }))
+      device:emit_event_for_endpoint(device:get_field(common_utils.ENERGY_MANAGEMENT_ENDPOINT), capabilities.energyMeter.energy({ value = watt_hour_value, unit = "Wh" }))
     end
   end
 end
@@ -1453,39 +705,6 @@ local function energy_report_handler_factory(is_cumulative_report)
   end
 end
 
-local function initial_press_event_handler(driver, device, ib, response)
-  if get_field_for_endpoint(device, SUPPORTS_MULTI_PRESS, ib.endpoint_id) then
-    -- Receipt of an InitialPress event means we do not want to ignore the next MultiPressComplete event
-    -- or else we would potentially not create the expected button capability event
-    set_field_for_endpoint(device, IGNORE_NEXT_MPC, ib.endpoint_id, nil)
-  elseif get_field_for_endpoint(device, INITIAL_PRESS_ONLY, ib.endpoint_id) then
-    device:emit_event_for_endpoint(ib.endpoint_id, capabilities.button.button.pushed({state_change = true}))
-  elseif get_field_for_endpoint(device, EMULATE_HELD, ib.endpoint_id) then
-    -- if our button doesn't differentiate between short and long holds, do it in code by keeping track of the press down time
-    init_press(device, ib.endpoint_id)
-  end
-end
-
--- if the device distinguishes a long press event, it will always be a "held"
--- there's also a "long release" event, but this event is required to come first
-local function long_press_event_handler(driver, device, ib, response)
-  device:emit_event_for_endpoint(ib.endpoint_id, capabilities.button.button.held({state_change = true}))
-  if get_field_for_endpoint(device, SUPPORTS_MULTI_PRESS, ib.endpoint_id) then
-    -- Ignore the next MultiPressComplete event if it is sent as part of this "long press" event sequence
-    set_field_for_endpoint(device, IGNORE_NEXT_MPC, ib.endpoint_id, true)
-  end
-end
-
-local function short_release_event_handler(driver, device, ib, response)
-  if not get_field_for_endpoint(device, SUPPORTS_MULTI_PRESS, ib.endpoint_id) then
-    if get_field_for_endpoint(device, EMULATE_HELD, ib.endpoint_id) then
-      emulate_held_event(device, ib.endpoint_id)
-    else
-      device:emit_event_for_endpoint(ib.endpoint_id, capabilities.button.button.pushed({state_change = true}))
-    end
-  end
-end
-
 local function active_power_handler(driver, device, ib, response)
   if ib.data.value then
     local watt_value = ib.data.value / CONVERSION_CONST_MILLIWATT_TO_WATT
@@ -1493,7 +712,7 @@ local function active_power_handler(driver, device, ib, response)
       device:emit_event_for_endpoint(ib.endpoint_id, capabilities.powerMeter.power({ value = watt_value, unit = "W"}))
     else
       -- when energy management is defined in the root endpoint(0), replace it with the first switch endpoint and process it.
-      device:emit_event_for_endpoint(device:get_field(ENERGY_MANAGEMENT_ENDPOINT), capabilities.powerMeter.power({ value = watt_value, unit = "W"}))
+      device:emit_event_for_endpoint(device:get_field(common_utils.ENERGY_MANAGEMENT_ENDPOINT), capabilities.powerMeter.power({ value = watt_value, unit = "W"}))
     end
   end
 end
@@ -1510,28 +729,6 @@ local function valve_level_attr_handler(driver, device, ib, response)
   if ib.data.value then
     device:emit_event_for_endpoint(ib.endpoint_id, capabilities.level.level(ib.data.value))
   end
-end
-
-local function multi_press_complete_event_handler(driver, device, ib, response)
-  -- in the case of multiple button presses
-  -- emit number of times, multiple presses have been completed
-  if ib.data and not get_field_for_endpoint(device, IGNORE_NEXT_MPC, ib.endpoint_id) then
-    local press_value = ib.data.elements.total_number_of_presses_counted.value
-    --capability only supports up to 6 presses
-    if press_value < 7 then
-      local button_event = capabilities.button.button.pushed({state_change = true})
-      if press_value == 2 then
-        button_event = capabilities.button.button.double({state_change = true})
-      elseif press_value > 2 then
-        button_event = capabilities.button.button(string.format("pushed_%dx", press_value), {state_change = true})
-      end
-
-      device:emit_event_for_endpoint(ib.endpoint_id, button_event)
-    else
-      log.info(string.format("Number of presses (%d) not supported by capability", press_value))
-    end
-  end
-  set_field_for_endpoint(device, IGNORE_NEXT_MPC, ib.endpoint_id, nil)
 end
 
 local function battery_percent_remaining_attr_handler(driver, device, ib, response)
@@ -1556,60 +753,42 @@ local function power_source_attribute_list_handler(driver, device, ib, response)
     -- Re-profile the device if BatPercentRemaining (Attribute ID 0x0C) or
     -- BatChargeLevel (Attribute ID 0x0E) is present.
     if attr.value == 0x0C then
-      battery_attr_support = battery_support.BATTERY_PERCENTAGE
+      battery_attr_support = common_utils.battery_support.BATTERY_PERCENTAGE
       break
     elseif attr.value == 0x0E then
-      battery_attr_support = battery_support.BATTERY_LEVEL
+      battery_attr_support = common_utils.battery_support.BATTERY_LEVEL
       break
     end
   end
-  if supports_modular_profile(device) then
-    match_modular_profile(driver, device, battery_attr_support)
-  else
-    local profile_name
-    if battery_attr_support == battery_support.BATTERY_PERCENTAGE then
-      profile_name = "button-battery"
-    elseif battery_attr_support == battery_support.BATTERY_LEVEL then
-      profile_name = "button-batteryLevel"
-    end
-    local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
-    if #button_eps > 1 then
-      profile_name = string.format("%d-", #button_eps) .. profile_name
-    end
-    if device.manufacturer_info.vendor_id == AQARA_MANUFACTURER_ID and
-       device.manufacturer_info.product_id == AQARA_CLIMATE_SENSOR_W100_ID then
-      profile_name = profile_name .. "-temperature-humidity"
-    end
-    if profile_name then
-      device:try_update_metadata({ profile = profile_name })
-    end
+  local profile_name
+  if battery_attr_support == common_utils.battery_support.BATTERY_PERCENTAGE then
+    profile_name = "button-battery"
+  elseif battery_attr_support == common_utils.battery_support.BATTERY_LEVEL then
+    profile_name = "button-batteryLevel"
   end
-end
-
-local function max_press_handler(driver, device, ib, response)
-  local max = ib.data.value or 1 --get max number of presses
-  device.log.debug("Device supports "..max.." presses")
-  -- capability only supports up to 6 presses
-  if max > 6 then
-    log.info("Device supports more than 6 presses")
-    max = 6
+  local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
+  if #button_eps > 1 then
+    profile_name = string.format("%d-", #button_eps) .. profile_name
   end
-  local MSL = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH_LONG_PRESS})
-  local supportsHeld = tbl_contains(MSL, ib.endpoint_id)
-  local values = create_multi_press_values_list(max, supportsHeld)
-  device:emit_event_for_endpoint(ib.endpoint_id, capabilities.button.supportedButtonValues(values, {visibility = {displayed = false}}))
+  if device.manufacturer_info.vendor_id == AQARA_MANUFACTURER_ID and
+     device.manufacturer_info.product_id == AQARA_CLIMATE_SENSOR_W100_ID then
+    profile_name = profile_name .. "-temperature-humidity"
+  end
+  if profile_name then
+    device:try_update_metadata({ profile = profile_name })
+  end
 end
 
 local function info_changed(driver, device, event, args)
-  if device:get_field(SUPPORTED_COMPONENT_CAPABILITIES) then
+  if device:get_field(modular_profiles_utils.SUPPORTED_COMPONENT_CAPABILITIES) then
     --re-up subscription with new capabilities using the modular supports_capability override
-    device:extend_device("supports_capability_by_id", supports_capability_by_id_modular)
+    device:extend_device("supports_capability_by_id", modular_profiles_utils.supports_capability_by_id_modular)
   end
   if device.profile.id ~= args.old_st_store.profile.id then
     device:subscribe()
     local button_eps = device:get_endpoints(clusters.Switch.ID, {feature_bitmap=clusters.Switch.types.SwitchFeature.MOMENTARY_SWITCH})
     if #button_eps > 0 and device.network_type == device_lib.NETWORK_TYPE_MATTER then
-      configure_buttons(device)
+      button_utils.configure_buttons(device)
     end
   end
 end
@@ -1650,9 +829,9 @@ local temp_attr_handler_factory = function(minOrMax)
     end
     local temp = ib.data.value / 100.0
     local unit = "C"
-    set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..minOrMax, ib.endpoint_id, temp)
-    local min = get_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MIN, ib.endpoint_id)
-    local max = get_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MAX, ib.endpoint_id)
+    common_utils.set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..minOrMax, ib.endpoint_id, temp)
+    local min = common_utils.get_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MIN, ib.endpoint_id)
+    local max = common_utils.get_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MAX, ib.endpoint_id)
     if min ~= nil and max ~= nil then
       if min < max then
         -- Only emit the capability for RPC version >= 5 (unit conversion for
@@ -1660,8 +839,8 @@ local temp_attr_handler_factory = function(minOrMax)
         if version.rpc >= 5 then
           device:emit_event_for_endpoint(ib.endpoint_id, capabilities.temperatureMeasurement.temperatureRange({ value = { minimum = min, maximum = max }, unit = unit }))
         end
-        set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MIN, ib.endpoint_id, nil)
-        set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MAX, ib.endpoint_id, nil)
+        common_utils.set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MIN, ib.endpoint_id, nil)
+        common_utils.set_field_for_endpoint(device, TEMP_BOUND_RECEIVED..TEMP_MAX, ib.endpoint_id, nil)
       else
         device.log.warn_with({hub_logs = true}, string.format("Device reported a min temperature %d that is not lower than the reported max temperature %d", min, max))
       end
@@ -1797,7 +976,7 @@ local matter_driver_template = {
         [clusters.PowerSource.attributes.BatPercentRemaining.ID] = battery_percent_remaining_attr_handler,
       },
       [clusters.Switch.ID] = {
-        [clusters.Switch.attributes.MultiPressMax.ID] = max_press_handler
+        [clusters.Switch.attributes.MultiPressMax.ID] = button_utils.max_press_handler
       },
       [clusters.RelativeHumidityMeasurement.ID] = {
         [clusters.RelativeHumidityMeasurement.attributes.MeasuredValue.ID] = humidity_attr_handler
@@ -1815,10 +994,10 @@ local matter_driver_template = {
     },
     event = {
       [clusters.Switch.ID] = {
-        [clusters.Switch.events.InitialPress.ID] = initial_press_event_handler,
-        [clusters.Switch.events.LongPress.ID] = long_press_event_handler,
-        [clusters.Switch.events.ShortRelease.ID] = short_release_event_handler,
-        [clusters.Switch.events.MultiPressComplete.ID] = multi_press_complete_event_handler
+        [clusters.Switch.events.InitialPress.ID] = button_utils.initial_press_event_handler,
+        [clusters.Switch.events.LongPress.ID] = button_utils.long_press_event_handler,
+        [clusters.Switch.events.ShortRelease.ID] = button_utils.short_release_event_handler,
+        [clusters.Switch.events.MultiPressComplete.ID] = button_utils.multi_press_complete_event_handler
       }
     },
     fallback = matter_handler,
@@ -1923,41 +1102,14 @@ local matter_driver_template = {
       [capabilities.fanSpeedPercent.commands.setPercent.NAME] = set_fan_speed_percent
     }
   },
-  supported_capabilities = {
-    capabilities.switch,
-    capabilities.switchLevel,
-    capabilities.colorControl,
-    capabilities.colorTemperature,
-    capabilities.level,
-    capabilities.motionSensor,
-    capabilities.illuminanceMeasurement,
-    capabilities.powerMeter,
-    capabilities.energyMeter,
-    capabilities.powerConsumptionReport,
-    capabilities.valve,
-    capabilities.button,
-    capabilities.battery,
-    capabilities.batteryLevel,
-    capabilities.temperatureMeasurement,
-    capabilities.relativeHumidityMeasurement,
-    capabilities.fanMode,
-    capabilities.fanSpeedPercent
-  },
+  supported_capabilities = common_utils.supported_capabilities,
   sub_drivers = {
-    require("eve-energy"),
     require("aqara-cube"),
+    require("eve-energy"),
+    require("modular-profiles"),
     require("third-reality-mk1")
   }
 }
-
-function detect_matter_thing(device)
-  for _, capability in ipairs(matter_driver_template.supported_capabilities) do
-    if device:supports_capability(capability) then
-      return false
-    end
-  end
-  return device:supports_capability(capabilities.refresh)
-end
 
 local matter_driver = MatterDriver("matter-switch", matter_driver_template)
 log.info_with({hub_logs=true}, string.format("Starting %s driver, with dispatcher: %s", matter_driver.NAME, matter_driver.matter_dispatcher))
